Activity structure / Clear actoken
