+++ teach/src/main/java/com/u/teach/activity/register/RegisterActivity.java
@@ -1,0 +1,10 @@
+package com.u.teach.activity.register;
+
+import com.u.teach.activity.DefaultActivity;
+
+/**
+ * Created by saguilera on 1/11/17.
+ */
+
+public class RegisterActivity extends DefaultActivity {
+}
